refactor(marketing): fixed-position linked badge, shared consent helper
