+++ src/components/MarketingView/styles.module.css
@@ -164,6 +164,12 @@
   color: var(--theme-elevation-800, #333);
 }
 
+.badgeMuted {
+  background: transparent;
+  color: var(--theme-elevation-400, #999);
+  border: 1px dashed var(--theme-elevation-150, #ececec);
+}
+
 /* Pagination */
 .pagination {
   display: flex;
